# Server/user-service/src/main/resources/specification/api-spec-users.yaml
openapi: 3.0.3
info:
  title: User Service API
  version: 1.0.0
  description: API for managing users in the SaaS platform

servers:
  - url: /api/users
    description: User Service API Base Path

paths:
  /users/getUsers:
    get:
      summary: Get All Users
      description: Retrieve a list of all users.
      operationId: getAllUsers
      tags:
        - User
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'

  /users/getUser/{id}:
    get:
      summary: Get User by ID
      description: Retrieve a single user by their unique ID.
      operationId: getUserById
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found

  /users/updateUser/{id}:
    put:
      summary: Update User by ID
      description: Update an existing user's information.
      operationId: updateUserById
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
  /users/deleteUser/{id}:
    delete:
      summary: Delete User by ID
      description: Delete a user by their unique ID.
      operationId: deleteUserById
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found

  /users/createUser:
    post:
      summary: Create a New User
      description: Create a new user with the provided information.
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid input

components:
  schemas:
    UserDTO:
      type: object
      properties:
        userId:
          type: integer
          description: Unique identifier for the user
        email:
          type: string
          format: email
          description: User's email address
        passwordHash:
          type: string
          description: Hashed password for secure storage
        createdAt:
          type: string
          format: date-time
          description: User creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        status:
          type: string
          enum: [ active, suspended, deleted ]
          description: Current status of the user
        role:
          type: integer
          description: User Role Id

    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: Plain text password (to be hashed server-side)

    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        status:
          type: string
          enum: [ active, suspended, deleted ]
          description: User's current status
        passwordHash:
          type: string
          description: Hashed password for secure storage
        role:
          type: integer
          description: User Role Id